re-organize dictator chain extension into re-usable types
