+++ contracts/dictator/Cargo.toml
@@ -10,6 +10,7 @@ ink_env = { version = "4.2.0", default-features = false }
 
 scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
 scale-info = { version = "2.3", default-features = false, features = ["derive"], optional = true }
+game-theory = { path = "../../", default-features = false }
 
 [dev-dependencies]
 ink_e2e = "4.0.1"
@@ -23,6 +24,7 @@ std = [
     "ink/std",
     "scale/std",
     "scale-info/std",
+    "game-theory/std",
 ]
 ink-as-dependency = []
 e2e-tests = []
