+++ logics/traits/types.rs
@@ -1,8 +1,11 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+
 use openbrush::contracts::access_control::AccessControlError;
 use openbrush::traits::{AccountId, Hash};
 use scale::{Decode, Encode};
 use ink::prelude::vec::Vec;
 use ink::storage::traits::StorageLayout;
+use ink::env::Environment;
 
 #[derive(Encode, Decode, Debug, PartialEq, Eq)]
 #[cfg_attr(feature = "std", derive(scale_info::TypeInfo))]
@@ -104,4 +107,43 @@ pub struct GameConfigs {
     pub is_rounds_based: bool,
 }
 
+#[derive(Debug, Copy, Clone, PartialEq, Eq, scale::Encode, scale::Decode)]
+#[cfg_attr(feature = "std", derive(scale_info::TypeInfo))]
+pub enum RandomReadErr {
+    FailGetRandomSource,
+}
 
+impl ink::env::chain_extension::FromStatusCode for RandomReadErr {
+    fn from_status_code(status_code: u32) -> Result<(), Self> {
+        match status_code {
+            0 => Ok(()),
+            1 => Err(Self::FailGetRandomSource),
+            _ => panic!("encountered unknown status code"),
+        }
+    }
+}
+
+#[ink::chain_extension]
+pub trait FetchRandom {
+    type ErrorCode = RandomReadErr;
+
+    #[ink(extension = 12)]
+    fn fetch_random(subject: [u8; 32]) -> [u8; 32];
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+#[cfg_attr(feature = "std", derive(scale_info::TypeInfo))]
+pub enum CustomEnvironment {}
+
+impl Environment for CustomEnvironment {
+    const MAX_EVENT_TOPICS: usize =
+        <ink::env::DefaultEnvironment as Environment>::MAX_EVENT_TOPICS;
+
+    type AccountId = <ink::env::DefaultEnvironment as Environment>::AccountId;
+    type Balance = <ink::env::DefaultEnvironment as Environment>::Balance;
+    type Hash = <ink::env::DefaultEnvironment as Environment>::Hash;
+    type BlockNumber = <ink::env::DefaultEnvironment as Environment>::BlockNumber;
+    type Timestamp = <ink::env::DefaultEnvironment as Environment>::Timestamp;
+
+    type ChainExtension = FetchRandom;
+}
